requirements.txt
# Core
numpy
pandas
scipy

# Visualization
matplotlib
seaborn
plotly

# Machine Learning
scikit-learn
xgboost

# Dashboard
streamlit

# Database
sqlalchemy

# Data Ingestion
kagglehub

# Utilities
pyyaml
joblib
tqdm

# Optional but recommended
jupyter
ipykernel